requirements.txt
aiohttp
aiohttp[speedups]
asyncpg
SQLAlchemy
python-dotenv
bcrypt
pydantic
